+++ src/functionalTest/kotlin/com/cmgapps/license/LicensePluginJavaMultiProjectShould.kt
@@ -163,10 +163,10 @@ class LicensePluginJavaMultiProjectShould {
                     "</head>" +
                     "<body>" +
                     "<h3>Notice for packages:</h3>" +
+                    "<ul><li>Fake dependency name</li></ul><div class=\"license\"><p>Some license</p><a href=\"http://website.tld/\">http://website.tld/</a></div>" +
                     "<ul><li>Retrofit</li></ul><pre>" +
                     getFileContent("apache-2.0.txt") +
                     "</pre>" +
-                    "<ul><li>Fake dependency name</li></ul><div class=\"license\"><p>Some license</p><a href=\"http://website.tld/\">http://website.tld/</a></div>" +
                     "</body>" +
                     "</html>"
             )