+++ src/functionalTest/kotlin/com/cmgapps/license/LicensePluginMultiplatformShould.kt
@@ -164,13 +164,13 @@ class LicensePluginMultiplatformShould {
             File("$reportFolder/$taskName/licenses.txt").readText().trim(),
             `is`(
                 "Licenses\n" +
-                    "├─ com.google.firebase:firebase-core:10.0.1\n" +
+                    "├─ Fake dependency name:1.0.0\n" +
+                    "│  ├─ License: Some license\n" +
+                    "│  └─ URL: http://website.tld/\n" +
                     "├─ Retrofit:2.3.0\n" +
                     "│  ├─ License: Apache 2.0\n" +
                     "│  └─ URL: http://www.apache.org/licenses/LICENSE-2.0.txt\n" +
-                    "└─ Fake dependency name:1.0.0\n" +
-                    "   ├─ License: Some license\n" +
-                    "   └─ URL: http://website.tld/"
+                    "└─ com.google.firebase:firebase-core:10.0.1"
             )
         )
     }
@@ -230,10 +230,10 @@ class LicensePluginMultiplatformShould {
             File("$reportFolder/$taskName/licenses.txt").readText().trim(),
             `is`(
                 "Licenses\n" +
-                    "├─ com.google.firebase:firebase-core:10.0.1\n" +
-                    "└─ Retrofit:2.3.0\n" +
-                    "   ├─ License: Apache 2.0\n" +
-                    "   └─ URL: http://www.apache.org/licenses/LICENSE-2.0.txt"
+                    "├─ Retrofit:2.3.0\n" +
+                    "│  ├─ License: Apache 2.0\n" +
+                    "│  └─ URL: http://www.apache.org/licenses/LICENSE-2.0.txt\n" +
+                    "└─ com.google.firebase:firebase-core:10.0.1"
             )
         )
     }
@@ -282,10 +282,10 @@ class LicensePluginMultiplatformShould {
             File("$reportFolder/$taskName/licenses.txt").readText().trim(),
             `is`(
                 "Licenses\n" +
-                    "├─ com.google.firebase:firebase-core:10.0.1\n" +
-                    "└─ Fake dependency name:1.0.0\n" +
-                    "   ├─ License: Some license\n" +
-                    "   └─ URL: http://website.tld/"
+                    "├─ Fake dependency name:1.0.0\n" +
+                    "│  ├─ License: Some license\n" +
+                    "│  └─ URL: http://website.tld/\n" +
+                    "└─ com.google.firebase:firebase-core:10.0.1"
             )
         )
     }
@@ -334,10 +334,10 @@ class LicensePluginMultiplatformShould {
             File("$reportFolder/$taskName/licenses.txt").readText().trim(),
             `is`(
                 "Licenses\n" +
-                    "├─ com.google.firebase:firebase-core:10.0.1\n" +
-                    "└─ Fake dependency name:1.0.0\n" +
-                    "   ├─ License: Some license\n" +
-                    "   └─ URL: http://website.tld/"
+                    "├─ Fake dependency name:1.0.0\n" +
+                    "│  ├─ License: Some license\n" +
+                    "│  └─ URL: http://website.tld/\n" +
+                    "└─ com.google.firebase:firebase-core:10.0.1"
             )
         )
     }