+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 
 ### Changed
-- sort dependencies by name and version
+- Sort dependencies by name and version
 
 ### Deprecated